+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+# This is a generated file; do not edit or check into version control.
+export "FLUTTER_ROOT=/home/sahdeep/flutter"
+export "FLUTTER_APPLICATION_PATH=/home/sahdeep/StudioProjects/liquid_swipe_flutter/example"
+export "FLUTTER_TARGET=lib/main.dart"
+export "FLUTTER_BUILD_DIR=build"
+export "SYMROOT=${SOURCE_ROOT}/../build/ios"
+export "FLUTTER_FRAMEWORK_DIR=/home/sahdeep/flutter/bin/cache/artifacts/engine/ios"