+++ lib/Animation_Gesture/page_reveal.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:liquid_swipe/Clippers/CircularWave.dart';
+import 'package:liquid_swipe/Constants/Helpers.dart';
 
-import '../Constants/constants.dart';
-import '../WaveLayer.dart';
+import '../Clippers/WaveLayer.dart';
 
 /// This class reveals the next page in the liquid wave form.
 
@@ -9,24 +10,54 @@ class PageReveal extends StatelessWidget {
   final double revealPercent;
   final Widget child;
   final SlideDirection slideDirection;
+  final double iconPosition;
+  final WaveType waveType;
+  final double vertReveal;
 
   //Constructor
-  PageReveal({
-    this.revealPercent,
+  PageReveal({this.revealPercent,
     this.child,
     this.slideDirection,
-  });
+    this.iconPosition,
+    this.waveType, this.vertReveal });
 
   @override
   Widget build(BuildContext context) {
     //ClipPath clips our Container (page) with clipper based on path..
-    return new ClipPath(
-      clipper: new WaveLayer(
-          revealPercent: slideDirection == SlideDirection.leftToRight
-              ? 1.0 - revealPercent
-              : revealPercent,
-          slideDirection: slideDirection),
-      child: child,
-    );
+    switch (waveType) {
+      case WaveType.liquidReveal:
+        return ClipPath(
+          clipper: WaveLayer(
+              revealPercent: slideDirection == SlideDirection.leftToRight
+                  ? 1.0 - revealPercent
+                  : revealPercent,
+              slideDirection: slideDirection,
+              iconPosition: iconPosition,
+              verReveal: vertReveal),
+          child: child,
+        );
+        break;
+      case WaveType.circularReveal:
+        return ClipPath(
+          clipper: CircularWave(iconPosition,
+              revealPercent: slideDirection == SlideDirection.leftToRight
+                  ? 1.0 - revealPercent
+                  : revealPercent),
+          child: child,
+        );
+        break;
+      default:
+        return ClipPath(
+          clipper: WaveLayer(
+              revealPercent: slideDirection == SlideDirection.leftToRight
+                  ? 1.0 - revealPercent
+                  : revealPercent,
+              slideDirection: slideDirection,
+              iconPosition: iconPosition,
+              verReveal: vertReveal),
+          child: child,
+        );
+        break;
+    }
   }
 }