+++ lib/liquid_swipe.dart
@@ -4,12 +4,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:liquid_swipe/Animation_Gesture/animated_page_dragger.dart';
 import 'package:liquid_swipe/Animation_Gesture/page_dragger.dart';
-import 'package:liquid_swipe/Constants/constants.dart';
 import 'package:liquid_swipe/page.dart';
 
 import 'Animation_Gesture/page_reveal.dart';
+import 'Constants/Helpers.dart';
 
 final key = new GlobalKey<_LiquidSwipe>();
+
+typedef OnPageChangeCallback = void Function(int activePageIndex);
 
 class LiquidSwipe extends StatefulWidget {
   final List<Container> pages;
@@ -17,6 +19,11 @@ class LiquidSwipe extends StatefulWidget {
   final int initialPage;
   final bool enableSlideIcon;
   final Widget slideIconWidget;
+  final double positionSlideIcon;
+  final bool enableLoop;
+  final WaveType waveType;
+
+  final OnPageChangeCallback onPageChangeCallback;
 
   const LiquidSwipe({
     Key key,
@@ -25,11 +32,17 @@ class LiquidSwipe extends StatefulWidget {
     this.initialPage = 0,
     this.enableSlideIcon = false,
     this.slideIconWidget = const Icon(Icons.arrow_back_ios),
+    this.positionSlideIcon = 0.54,
+    this.enableLoop = true,
+    this.waveType = WaveType.liquidReveal,
+    this.onPageChangeCallback,
   })  : assert(pages != null),
+        assert(onPageChangeCallback != null),
         assert(fullTransitionValue != null),
         assert(initialPage != null &&
             initialPage >= 0 &&
             initialPage < pages.length),
+        assert(positionSlideIcon >= -1 && positionSlideIcon <= 1),
         super(key: key);
 
   @override
@@ -39,11 +52,12 @@ class LiquidSwipe extends StatefulWidget {
 class SlideUpdate {
   final UpdateType updateType;
   final SlideDirection direction;
-  final double slidePercent;
+  final double slidePercentHor, slidePercentVer;
 
   SlideUpdate(
     this.direction,
-    this.slidePercent,
+      this.slidePercentHor,
+      this.slidePercentVer,
     this.updateType,
   );
 }
@@ -59,11 +73,19 @@ class _LiquidSwipe extends State<LiquidSwipe> with TickerProviderStateMixin {
   int activePageIndex = 0; //active page index
   int nextPageIndex = 0; //next page index
   SlideDirection slideDirection = SlideDirection.none; //slide direction
-  double slidePercent = 0.0; //slide percentage (0.0 to 1.0)
+  double slidePercentHor,
+      slidePercentVer = 0.0; //slide percentage (0.0 to 1.0)
   StreamSubscription<SlideUpdate> slideUpdateStream$;
+
+  set setActivePageIndex(int value) {
+    activePageIndex = value;
+
+    widget.onPageChangeCallback(activePageIndex);
+  }
 
   @override
   void initState() {
+    slidePercentHor = slidePercentVer = 0;
     activePageIndex = widget.initialPage;
     nextPageIndex = widget.initialPage;
     //Stream Controller initialization
@@ -76,29 +98,45 @@ class _LiquidSwipe extends State<LiquidSwipe> with TickerProviderStateMixin {
         //if the user is dragging then
         if (event.updateType == UpdateType.dragging) {
           slideDirection = event.direction;
-          slidePercent = event.slidePercent;
+          slidePercentHor = event.slidePercentHor;
+          slidePercentVer = event.slidePercentVer;
 
-          //conditions on slide direction
-          if (slideDirection == SlideDirection.leftToRight) {
-            nextPageIndex = activePageIndex - 1;
-          } else if (slideDirection == SlideDirection.rightToLeft) {
-            nextPageIndex = activePageIndex + 1;
-          } else {
-            nextPageIndex = activePageIndex;
-          }
           // making pages to be in loop
-          if (nextPageIndex > widget.pages.length - 1)
-            nextPageIndex = 0;
-          else if (nextPageIndex < 0) nextPageIndex = widget.pages.length - 1;
+          if (widget.enableLoop) {
+            //conditions on slide direction
+            if (slideDirection == SlideDirection.leftToRight) {
+              nextPageIndex = activePageIndex - 1;
+            } else if (slideDirection == SlideDirection.rightToLeft) {
+              nextPageIndex = activePageIndex + 1;
+            } else {
+              nextPageIndex = activePageIndex;
+            }
+
+            if (nextPageIndex > widget.pages.length - 1)
+              nextPageIndex = 0;
+            else if (nextPageIndex < 0) nextPageIndex = widget.pages.length - 1;
+          } else {
+            //conditions on slide direction
+            if (slideDirection == SlideDirection.leftToRight &&
+                activePageIndex != 0) {
+              nextPageIndex = activePageIndex - 1;
+            } else if (slideDirection == SlideDirection.rightToLeft &&
+                activePageIndex != widget.pages.length - 1) {
+              nextPageIndex = activePageIndex + 1;
+            } else {
+              nextPageIndex = activePageIndex;
+            }
+          }
         }
         //if the user has done dragging
         else if (event.updateType == UpdateType.doneDragging) {
           // slidepercent > 0.2 so that it wont reveal itself unless this condition is true
-          if (slidePercent > 0.2) {
+          if (slidePercentHor > 0.2) {
             animatedPageDragger = AnimatedPageDragger(
               slideDirection: slideDirection,
               transitionGoal: TransitionGoal.open,
-              slidePercent: slidePercent,
+              slidePercentHor: slidePercentHor,
+              slidePercentVer: slidePercentVer,
               slideUpdateStream: slideUpdateStream,
               vsync: this,
             );
@@ -106,16 +144,13 @@ class _LiquidSwipe extends State<LiquidSwipe> with TickerProviderStateMixin {
             animatedPageDragger = AnimatedPageDragger(
               slideDirection: slideDirection,
               transitionGoal: TransitionGoal.close,
-              slidePercent: slidePercent,
+              slidePercentHor: slidePercentHor,
+              slidePercentVer: slidePercentVer,
               slideUpdateStream: slideUpdateStream,
               vsync: this,
             );
 
             nextPageIndex = activePageIndex;
-            //to continue in the loop of pages
-            if (nextPageIndex > widget.pages.length - 1)
-              nextPageIndex = 0;
-            else if (nextPageIndex < 0) nextPageIndex = widget.pages.length - 1;
           }
           //Run the animation
           animatedPageDragger.run();
@@ -123,13 +158,16 @@ class _LiquidSwipe extends State<LiquidSwipe> with TickerProviderStateMixin {
         //when animating
         else if (event.updateType == UpdateType.animating) {
           slideDirection = event.direction;
-          slidePercent = event.slidePercent;
+          slidePercentHor = event.slidePercentHor;
+          slidePercentVer = event.slidePercentVer;
         }
         //done animating
         else if (event.updateType == UpdateType.doneAnimating) {
           activePageIndex = nextPageIndex;
+          widget.onPageChangeCallback(activePageIndex);
           slideDirection = SlideDirection.none;
-          slidePercent = 0.5;
+          slidePercentHor = 0.5;
+          slidePercentVer = widget.positionSlideIcon;
         }
       });
     });
@@ -156,18 +194,20 @@ class _LiquidSwipe extends State<LiquidSwipe> with TickerProviderStateMixin {
             pageView: slideDirection == SlideDirection.leftToRight
                 ? pages[activePageIndex]
                 : pages[nextPageIndex],
-            percentVisible: 1.0,
           ),
           //Pages
           PageReveal(
             //next page reveal
-            revealPercent: slidePercent,
+            revealPercent: slidePercentHor,
             child: Page(
                 pageView: slideDirection == SlideDirection.leftToRight
                     ? pages[nextPageIndex]
-                    : pages[activePageIndex],
-                percentVisible: slidePercent),
+                    : pages[activePageIndex]),
             slideDirection: slideDirection,
+            iconPosition: widget.positionSlideIcon,
+
+            waveType: widget.waveType,
+            vertReveal: slidePercentVer,
           ),
           PageDragger(
             //Used for gesture control
@@ -175,16 +215,10 @@ class _LiquidSwipe extends State<LiquidSwipe> with TickerProviderStateMixin {
             slideUpdateStream: this.slideUpdateStream,
             enableSlideIcon: widget.enableSlideIcon,
             slideIconWidget: widget.slideIconWidget,
+            iconPosition: widget.positionSlideIcon,
           ), //PageDragger
         ], //Widget
       ), //Stack
     ); //Scaffold
-  }
-
-  next() {
-    _LiquidSwipe().setState(() {
-      activePageIndex += 1;
-      nextPageIndex = activePageIndex + 1;
-    });
   }
 }