+++ CHANGELOG.md
@@ -1,7 +1,16 @@
+#ChangeLog
+
+## 1.2.0
+* CustomIcon [5cbd32c583126e14d135ad3e9d6e6ea8221edb3d](https://github.com/iamSahdeep/liquid_swipe_flutter/commit/5cbd32c583126e14d135ad3e9d6e6ea8221edb3d)
+* Enable page recurrence [fdbcfc1e14e148201685fc401ea2505cc26a5cd4](https://github.com/iamSahdeep/liquid_swipe_flutter/commit/fdbcfc1e14e148201685fc401ea2505cc26a5cd4)
+* New Clipper : CircularReveal [ad844b2e9c0cdaa124e2a2a76caad5cc35dce67b](https://github.com/iamSahdeep/liquid_swipe_flutter/commit/ad844b2e9c0cdaa124e2a2a76caad5cc35dce67b)
+
+
 ## 1.1.0
 * Added Initial page attribute. [cda6de80113502f17da8085d95c5eaea543f42b7](https://github.com/iamSahdeep/liquid_swipe_flutter/commit/cda6de80113502f17da8085d95c5eaea543f42b7)
 * Added Slide Icon on pages. [32680cdc4571a705b4e7f11467b450e37016807d](https://github.com/iamSahdeep/liquid_swipe_flutter/commit/32680cdc4571a705b4e7f11467b450e37016807d)
 * Other Minor Fixes. [52d159403525261dbfaa99a52af718a1ddc8274f](https://github.com/iamSahdeep/liquid_swipe_flutter/commit/52d159403525261dbfaa99a52af718a1ddc8274f) 
+
 ## 1.0.2
 
 * Minor spell changes